Fix filtering on server for `contains`
The server side filtering for checking that the body contains a certain string was looking for the Todo to have a field `contains` rather than `body`. This was causing problems since none of the todos have a field called `contains`. This probably happened in the first place since the user example used `category` in both the parameter and the field.

Co-authored-by: Will Marsan <marsa027@morris.umn.edu>